requirements.txt
pandas==2.1.0
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
scipy==1.11.2
scikit-learn==1.3.0
jupyter==1.0.0
plotly==5.16.1
requests==2.31.0
python-dotenv==1.0.0
eia-python==1.25
kaggle==1.5.16 